Allocate the framebuffer on a shared memory segment.
This is a quick modification to Chocolate Doom that exposes the
framebuffer to external applications through a System V shared
memory segment. Applications wishing to retrieve the most recent
snapshot of the game screen can do so by attaching to shmid '666'
(because Doom!) and read the data straight from that area.

The first 256*3 bytes hold the RGB lookup table, while the bytes
following that are the actual framebuffer (320*200 bytes of data).